Organize event deletion code for easier management
